requirements.txt
pip
autopep8
Flask==3.0.3
gunicorn==22.0.0
Werkzeug==3.0.6
Flask-JWT-Extended==4.7.1
Flask-SQLAlchemy==3.1.1
PyMySQL==1.1.1
google-cloud-storage==3.1.0
prometheus-client==0.22.0
locust==2.37.10
snakeviz==2.2.2

Flask-Migrate==4.0.5
Authlib==1.3.0
requests==2.32.2
python-dotenv==1.0.1
cryptography==42.0.0
mysql-connector-python==9.3.0